Disable the redirection due to DNS/openshift interaction restriction
Bind refuse to have a CNAME at the root of the domain ( ie, projectatomic.io ),
and openshift will not work if we do not give a CNAME, as the node can be migrated,
etc. If we want to avoid www.projectatomic.io, we need to rent a server and/or use
another hosting.