Add virt-manager cloud-init to Quickstart
The current quickstart seems a bit confusing when it comes to the login part. At first glance it appears that the instructions for virt-manager are missing but in fact it just links to a blog (that might go away at some point) and proceeds to explain the process for VirtualBox.

I've added some information on how to generate the image - it's the same for both platforms, and also how to make sure your VM will pick it up in virt-manager.