remove links to unsupported documentation
The Getting Started Guide for setting up an Atomic Cluster and the
docs around Kubernetes are old, unmaintained, and hard to provide
support for.

Let's stop pointing users at these docs and try to find a better way
for users to experiment with kubernetes, etc.

This is not an exhaustive removal of the links to both guides, but the
primary entrypoints that I found.