Modified Team Silverblue URL
The canonical URL for the Silverblue project appears to be https://silverblue.fedoraproject.org, not https://teamsilverblue.org/. While https://teamsilverblue.org/ appears to have redirected to the proper site in the past it currently does not. Also, the ownership of teamsilverblue.org is unclear but it does not appear to be a Fedora Project domain.